+++ Sources/CompassUtils/Files.swift
@@ -1,0 +1,24 @@
+import Foundation
+
+public func getCurrentWorkingDirectory() -> String {
+    return FileManager.default.currentDirectoryPath
+}
+
+public func isDirectoryEmpty(atPath path: String) -> Bool {
+    do {
+        let contents = try FileManager.default.contentsOfDirectory(atPath: path)
+        return contents.isEmpty
+    } catch {
+        return false
+    }
+}
+
+public func isDirectoryExists(atPath path: String) -> Bool {
+    var isDirectory: ObjCBool = false
+    FileManager.default.fileExists(atPath: path, isDirectory: &isDirectory)
+    return isDirectory.boolValue
+}
+
+public func isAbsolutePath(_ path: String) -> Bool {
+    return path.hasPrefix("/")
+}