+++ CHANGELOG.md
@@ -1,3 +1,15 @@
+# Version 0.0.4
+
+Setup init command to scaffold new Sailor projects. Followed [https://www.polpiella.dev/how-to-make-an-interactive-picker-for-a-swift-command-line-tool]this tutorial for the interactive picker.
+
+# Version 0.0.3
+
+Basic CLI architecture, filesystem and git helpers.
+
+# Version 0.0.2
+
+Uploaded to brew at []. Removed Carton as a dependency of CLI.
+
 # Version 0.0.1
 
 Attempting to get a swift CLI uploaded to brew.